use interactive mode in build-ios
